+++ responsibleai_vision/tests/test_feature_extractors.py
@@ -5,6 +5,7 @@ from common_vision_utils import (load_flowers_dataset, load_fridge_dataset,
                                  load_fridge_object_detection_dataset,
                                  load_imagenet_dataset)
 
+from responsibleai.feature_metadata import FeatureMetadata
 from responsibleai_vision.common.constants import (ExtractedFeatures,
                                                    ImageColumns, ImageModes)
 from responsibleai_vision.utils.feature_extractors import extract_features
@@ -18,17 +19,23 @@ FRIDGE_METADATA_FEATURES = [
 
 
 def validate_extracted_features(extracted_features, feature_names,
-                                expected_feature_names, data):
+                                expected_feature_names, data,
+                                feature_metadata=None):
     assert len(extracted_features) == len(data)
     assert feature_names[0] == expected_feature_names[0]
     for i in range(1, len(feature_names)):
         assert feature_names[i] in expected_feature_names
     assert len(feature_names) == len(expected_feature_names)
     assert len(extracted_features[0]) == len(feature_names)
+    if feature_metadata is not None:
+        assert len(feature_metadata.categorical_features) <= len(feature_names)
+        for categorical_feature in feature_metadata.categorical_features:
+            assert categorical_feature in feature_names
 
 
-def extract_dataset_features(data):
-    return extract_features(data, ImageColumns.LABEL, ImageModes.RGB, None)
+def extract_dataset_features(data, feature_metadata=None):
+    return extract_features(data, ImageColumns.LABEL, ImageModes.RGB,
+                            feature_metadata=feature_metadata)
 
 
 class TestFeatureExtractors(object):
@@ -55,15 +62,21 @@ class TestFeatureExtractors(object):
 
     def test_extract_features_flowers_metadata(self):
         data = load_flowers_dataset(upscale=False)
-        extracted_features, feature_names = extract_dataset_features(data)
+        feature_metadata = FeatureMetadata()
+        extracted_features, feature_names = extract_dataset_features(
+            data, feature_metadata=feature_metadata)
         expected_feature_names = [MEAN_PIXEL_VALUE]
         validate_extracted_features(extracted_features, feature_names,
-                                    expected_feature_names, data)
+                                    expected_feature_names, data,
+                                    feature_metadata)
 
     def test_extract_features_mixed_exif_XPComment_metadata(self):
         data = load_fridge_dataset(add_extra_mixed_metadata=True)
-        extracted_features, feature_names = extract_dataset_features(data)
+        feature_metadata = FeatureMetadata()
+        extracted_features, feature_names = extract_dataset_features(
+            data, feature_metadata=feature_metadata)
         expected_feature_names = [MEAN_PIXEL_VALUE, 'XPComment']
         expected_feature_names += FRIDGE_METADATA_FEATURES
         validate_extracted_features(extracted_features, feature_names,
-                                    expected_feature_names, data)
+                                    expected_feature_names, data,
+                                    feature_metadata)